Added CDATA support and more verbose error messages.
The CDATA is parsed and then handled in the same way as text is. It has
its own constructor in `Content`, and callbacks in suitable places.

The verbose error messages are implemented by giving the `XenoParseError` a
`ByteString` payload. This makes finding the limits of the parser much
easier, and it was very useful when I implemented the CDATA support.